Update and fix to RGB Neopixel block
This is a fix and repair to the mistake I made 3 days ago.

Renamed name in definitions from `pixels` to `pixel`  to match the name and keep the name standard with other blocks.

Added setups - begin, clear, and show. these are clearing the pixels and turning all pixels off.

This has been tested in my private forked repo and works well.