Input text for project name added
I added an input text field in the top-title section and moved the save button to the top-title. I added some css lines to make it look like other fields and some javascript lines to capture the value of the text field and add it to the file that will be saved.