Get name when load a file & scroll bar for blocks
When load a project from computer it gets the name and display it in the project name text input. 
Note: when loading an example from the software it didn't get the name of the example. 

Also, I added a visible scroll bar in the blocks palette when the blocks overflow the screen height.